* changed focus and popup handling of COptionMenu. When the menu takes focus it doesn't popup anymore. This way you can build 		full keyboard accessible plug-ins. Now if the menu has focus and the return key is pressed the menu will popup. 		You may need to change some of your code, if you manually called menu->takeFocs() change it to menu->popup()
	* added keyboard hook for CFrame (see IKeyboardHook)
	* added keyboard handling to IDataBrowser
	* moved background drawing in CParamDisplay in its own method drawBack() and changed drawText() method by removing the newBack parameter

	* fix CScrollView::makeRectVisible()
	* fix crash in COptionMenu when calling getEntry(index) with index higher than number of items in menu
	* fix CViewContainer::addView (view, beforeView)

+++ branches/uidescription/vstgui/cfileselector.cpp
@@ -139,7 +139,7 @@ CNewFileSelector::~CNewFileSelector ()
 	setTitle (0);
 	setInitialDirectory (0);
 	setDefaultSaveName (0);
-	for (int i = 0; i < result.size (); i++)
+	for (size_t i = 0; i < result.size (); i++)
 		free (result[i]);
 }
 
@@ -256,7 +256,7 @@ int CNewFileSelector::getNumSelectedFiles () const
 }
 
 //-----------------------------------------------------------------------------
-const char* CNewFileSelector::getSelectedFile (int index) const
+const char* CNewFileSelector::getSelectedFile (unsigned int index) const
 {
 	if (index < result.size ())
 		return result[index];
